refactor(mcp): split streamable-http into handle-post/session modules
P2 shipped streamable-http.ts as a 745-LOC god-file with handlePost
at 349 LOC. The complexity gate blocked every edit. P3 splits it
along its natural responsibility seams while keeping the external
surface stable.

Layout:
  src/mcp/streamable-http.ts          — 11-line re-export shim so
    existing imports (tests, plugins, external code) keep working.
  src/mcp/streamable-http/session.ts  — types, constants, session
    state map, shared helpers (CORS, readBody, jsonResponse, origin
    + SSE accept detection).
  src/mcp/streamable-http/handle-post.ts — POST /mcp dispatch. Split
    into small helpers (preflight, readAndParse, validateSession,
    handleTaskMethod, handleNotification, handleAsync,
    handleInitialize, handleSyncResponse, runHandlerSafe) so no
    function exceeds 80 lines.
  src/mcp/streamable-http/index.ts    — routing + startStreamableHttp
    public entry; OPTIONS / OAuth / GET-health / GET-/mcp / DELETE
    all dispatched from a single route() helper.

Wire behavior unchanged. Every test under tests/unit/streamable-http
.test.ts (31 cases) passes without modification.

MN1 closeout: Handler is now sync-or-async with undefined for
notifications, matching what buildHandler in handler.ts actually
returns. The pre-P2 cast-adapt in server.ts is deleted.
